Now installs passmenu as well
Summary:
passmenu is a dmenu based helper for password-store. It's now installed
along with the binaries.
Being that it's a script included with password-store, it seemed like a
good idea, and is really helpful in i3.

Test Plan: I installed the built package and ran /usr/bin/passmenu. Works exactly as expected.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.solus-project.com/D987